requirements.txt
claude-agent-sdk>=0.1.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0
python-dotenv>=1.0.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0
anyio>=4.0.0
